start change build system to cmake. task file changed to c++.

+++ src/kernel/kmm.cpp
@@ -1,10 +1,10 @@
 #include "kmm.hpp"
 
-void* operator new[](u32 s) {
+void* operator new[](long unsigned int s) {
     return kmalloc(s);
 }
 
-void* operator new(u32 s) {
+void* operator new(long unsigned int s) {
     return kmalloc(s);
 }
 
@@ -12,6 +12,6 @@ void operator delete[](void* p) {
     return kfree(p);
 }
 
-void operator delete(void* p, u32 s) {
+void operator delete(void* p, long unsigned int s) {
     return kfree(p);
 }
