+++ src/kernel/kmm.hpp
@@ -4,10 +4,10 @@ extern "C" {
 #include "kmm.h"
 }
 
-void* operator new[](u32 s);
+void* operator new[](long unsigned int s);
 
-void* operator new(u32 s);
+void* operator new(long unsigned int s);
 
 void operator delete[](void* p);
 
-void operator delete(void* p, u32);
+void operator delete(void* p, long unsigned int);
